fix(db): migration 010 — corregir INT→BIGINT en acct, ops, rest_admin
Basado en grep exhaustivo de TODOS los 138 archivos SP (839 funciones).

## Funciones corregidas
### usp_acct.sql
- usp_acct_pos_getheader: p_sale_ticket_id INT → BIGINT
- usp_acct_pos_gettaxsummary: p_sale_ticket_id INT → BIGINT
- usp_acct_rest_getheader: p_order_ticket_id INT → BIGINT
- usp_acct_rest_gettaxsummary: p_order_ticket_id INT → BIGINT

### usp_ops.sql
- usp_pos_waitticket_create: p_customer_id INT → BIGINT
- usp_pos_waitticketline_insert: p_product_id INT → BIGINT
- usp_pos_saleticket_create: p_customer_id INT → BIGINT
- usp_pos_saleticketline_insert: p_product_id INT → BIGINT
- usp_rest_orderticketline_insert: p_product_id INT → BIGINT
- usp_rest_orderticketline_getbyid: "productId" INT → BIGINT (RETURNS TABLE)
- usp_bank_movement_create: "movementId" INT → BIGINT (RETURNS TABLE)

### usp_rest_admin.sql
- usp_rest_admin_compra_update: p_supplier_id INT → BIGINT
- usp_rest_admin_adjuststock: p_product_id INT → BIGINT

+++ web/api/sqlweb-pg/includes/sp/usp_acct.sql
@@ -441,7 +441,7 @@ $$;
 -- =============================================================================
 DROP FUNCTION IF EXISTS usp_acct_pos_getheader(INT) CASCADE;
 CREATE OR REPLACE FUNCTION usp_acct_pos_getheader(
-    p_sale_ticket_id INT
+    p_sale_ticket_id BIGINT
 )
 RETURNS TABLE(
     "id"          BIGINT,
@@ -478,7 +478,7 @@ $$;
 -- =============================================================================
 DROP FUNCTION IF EXISTS usp_acct_pos_gettaxsummary(INT) CASCADE;
 CREATE OR REPLACE FUNCTION usp_acct_pos_gettaxsummary(
-    p_sale_ticket_id INT
+    p_sale_ticket_id BIGINT
 )
 RETURNS TABLE(
     "taxRate"     NUMERIC,
@@ -506,7 +506,7 @@ $$;
 -- =============================================================================
 DROP FUNCTION IF EXISTS usp_acct_rest_getheader(INT) CASCADE;
 CREATE OR REPLACE FUNCTION usp_acct_rest_getheader(
-    p_order_ticket_id INT
+    p_order_ticket_id BIGINT
 )
 RETURNS TABLE(
     "id"          BIGINT,
@@ -536,7 +536,7 @@ $$;
 -- =============================================================================
 DROP FUNCTION IF EXISTS usp_acct_rest_gettaxsummary(INT) CASCADE;
 CREATE OR REPLACE FUNCTION usp_acct_rest_gettaxsummary(
-    p_order_ticket_id INT
+    p_order_ticket_id BIGINT
 )
 RETURNS TABLE(
     "taxRate"     NUMERIC,
